admin: extract clusters handler to separate class (#12884)
